+++ src/main/java/com/app/service/DonationService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.app.model.Donation;
 
 public interface DonationService {
-	public Donation findById(int id) throws Exception;
+	public Donation findById(Long id) throws Exception;
 	public void deleteById(String id) throws Exception;
 	public void insert(Donation donation) throws Exception;
 	public void update(Donation donation) throws Exception;
